general renamings to make clear that our new implementation is _the_ implementation
