+++ bench/syclblas/syclblas_benchmark_level2_jose.cpp
@@ -29,7 +29,7 @@
 
 using namespace blas;
 
-BENCHMARK_NAME_FORMAT(syclblas_level_2) {
+BENCHMARK_NAME_FORMAT(syclblas_level_2_jose) {
   std::ostringstream fname;
   fname << benchmark<>::typestr<ElemT>() << "_" << name() << "_"
         << std::get<0>(params) << "_" << std::get<1>(params) << "_"
@@ -37,7 +37,7 @@ BENCHMARK_NAME_FORMAT(syclblas_level_2) {
   return fname.str();
 }
 
-BENCHMARK(gemv, syclblas_level_2) {
+BENCHMARK(gemv, syclblas_level_2_jose) {
   using ScalarT = ElemT;
   using IndexType = unsigned int;
 
@@ -74,8 +74,8 @@ BENCHMARK(gemv, syclblas_level_2) {
 
   benchmark<>::flops_units_t flops =
       benchmark<>::measure(reps, n_fl_ops, [&]() {
-        auto event = _gemv(ex, *t_str, m, n, alpha, m_a_gpu, m, v_b_gpu, incX,
-                           beta, v_c_gpu, incY);
+        auto event = _gemv_jose(ex, *t_str, m, n, alpha, m_a_gpu, m, v_b_gpu,
+                                incX, beta, v_c_gpu, incY);
         ex.wait(event);
       });
